+++ pages/testing.html
@@ -1,7 +1,3 @@
 <h1>Testing Page</h1>
 
 <h2>ECDSA</h2>
-
-<script>
-    const ecdsa = require("ecdsa")
-</script>
